requirements.txt
langchain-core~=0.3.45
langchain-openai==0.3.2
langchain-chroma==0.2.1
openai==1.60.2
chromadb==0.6.3
youtube-transcript-api==0.6.2
google-api-python-client==2.130.0
pytube==15.0.0
sqlalchemy==2.0.30
isodate==0.6.1
python-dotenv==1.0.1
tiktoken==0.7.0
requests~=2.32.3
langchain-text-splitters~=0.3.6
langchain~=0.3.20